requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Shelve==0.1.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.4
pathspec==0.8.1
Pillow==8.0.1
pycparser==2.20
PyJWT==2.0.0
pymongo==3.11.2
regex==2020.11.13
requests==2.25.1
scikit-learn==0.24.0
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.22
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
websockets==8.1
Werkzeug==1.0.1
WTForms==2.3.3
flask-uploads==0.2.1
